requirements.txt
pandas>=2.0.0
lark>=1.1.0
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
google-generativeai>=0.3.0

